Added function to plot evolution of three-variable BA iterations to check convergence.
Added small value eps() to the marginals in three-variable BA and
one-variable BA to improve numerical stability when computing KL-
divergences and log-ratios (distributions must not have entries that
are exactly 0, which happens due to limited numerical precision).

Changed notation for single-variable functions to use 'w' instead
of 'o', both in the library and the notebooks.